Rename test_utils to utils.
